izmenen tip peremennyh

+++ LAB_2/Form1.cs
@@ -19,9 +19,9 @@ namespace LAB_2
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToDouble(textBox1.Text);
+            double b = Convert.ToDouble(textBox2.Text);
+            double c;
             c = a + b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Сложение";
@@ -29,9 +29,9 @@ namespace LAB_2
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToDouble(textBox1.Text);
+            double b = Convert.ToDouble(textBox2.Text);
+            double c;
             c = a - b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Вычитание";
@@ -39,9 +39,9 @@ namespace LAB_2
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToDouble(textBox1.Text);
+            double b = Convert.ToDouble(textBox2.Text);
+            double c;
             c = a * b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Умножение";
@@ -49,9 +49,9 @@ namespace LAB_2
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToDouble(textBox1.Text);
+            double b = Convert.ToDouble(textBox2.Text);
+            double c;
             c = a / b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Деление";
